Add NetworkPane
