+++ src/components/single-line-diagram.js
@@ -8,15 +8,16 @@
 import React, {useEffect} from 'react';
 import PropTypes from "prop-types";
 
+import {useDispatch} from "react-redux";
+
+import Box from '@material-ui/core/Box';
+import CloseIcon from '@material-ui/icons/Close';
+import IconButton from '@material-ui/core/IconButton';
 import Paper from "@material-ui/core/Paper";
 import {makeStyles} from "@material-ui/core/styles";
 import Typography from '@material-ui/core/Typography';
-import CloseIcon from '@material-ui/icons/Close';
-import Box from '@material-ui/core/Box';
-import IconButton from '@material-ui/core/IconButton';
 
 import {removeVoltageLevelDiagram} from '../redux/actions';
-import {useDispatch} from "react-redux";
 
 const useStyles = makeStyles(theme => ({
     diagram: {
@@ -53,7 +54,7 @@ const SingleLineDiagram = (props) => {
         <Paper elevation={1} variant='outlined' className={classes.diagram}>
             <Box display="flex" flexDirection="row">
                 <Box flexGrow={1}>
-                    <Typography>{props.diagramId}</Typography>
+                    <Typography>{props.diagramTitle}</Typography>
                 </Box>
                 <IconButton className={classes.close} onClick={() => dispatch(removeVoltageLevelDiagram())}>
                     <CloseIcon/>
@@ -65,7 +66,7 @@ const SingleLineDiagram = (props) => {
 };
 
 SingleLineDiagram.propTypes = {
-    diagramId: PropTypes.string.isRequired,
+    diagramTitle: PropTypes.string.isRequired,
     svg: PropTypes.string.isRequired
 };
