+++ src/components/top-bar.js
@@ -5,20 +5,24 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import {ReactComponent as PowsyblLogo} from "../images/powsybl_logo.svg";
-import Typography from "@material-ui/core/Typography";
+import React from "react";
+
+import {useDispatch, useSelector} from "react-redux";
+
 import {FormattedMessage} from "react-intl";
+
+import AppBar from "@material-ui/core/AppBar";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
-import Switch from "@material-ui/core/Switch";
 import IconButton from "@material-ui/core/IconButton";
+import {makeStyles} from "@material-ui/core/styles";
+import Switch from "@material-ui/core/Switch";
+import Toolbar from "@material-ui/core/Toolbar";
+import Typography from "@material-ui/core/Typography";
 import BrightnessLowIcon from "@material-ui/icons/BrightnessLow";
 import BrightnessHighIcon from "@material-ui/icons/BrightnessHigh";
-import React from "react";
+
 import {selectDarkTheme, toggleUseNameState} from "../redux/actions";
-import {useDispatch, useSelector} from "react-redux";
-import {makeStyles} from "@material-ui/core/styles";
+import {ReactComponent as PowsyblLogo} from "../images/powsybl_logo.svg";
 
 const useStyles = makeStyles(theme => ({
     appBar: {
